mt76: mt7996: Add external eeprom support
Add external eeprom support for upstream mt76 by the patch on mediatek feed.
Without this commit, mt7990 and mt7992 cannot read eeprom on chip.
Note that this commit doesn't implement related debugfs, which is applied in the orginal patch.
Pass the test with: AsiaRF AW7990-AED (mt7990) and AW7991-AE2 (mt7992)

As the original patch described:
Add external eeprom support
For mt7992 and mt7990, efuse mode is not supported due to the lack of
space in efuse.
So, an additional external eeprom is added for user to store their
golden eeprom.

Note that the FW currently has some issues with writing to the ext
eeprom, so the write back function of ext eeprom is not yet linked
to any command.
A write back command will be added once the FW fixes the issue.

Add ext eeprom write callback

Link: https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/mtk-openwrt-feeds/+/refs/heads/master/autobuild/unified/filogic/mac80211/24.10/files/package/kernel/mt76/patches/0033-mtk-mt76-mt7996-add-external-eeprom-support.patch
Signed-off-by: Elwin Huang <s09289728096@gmail.com>